refactor(packages/explorer/src/views/Mining/index.js packages/explorer/src/views/Mining/tokenMiner.j
Separating concerns
